refactor CalendarBundle part 1
